split index.d.ts into individual files
