+++ types/Modal.d.ts
@@ -1,0 +1,9 @@
+import { PopupProps, PopupContentProps } from "./Popup";
+
+export type ModalProps = PopupProps & { size?: "small" | "medium" | "large" };
+
+export const Modal: React.FC<ModalProps> & {
+  Header: React.FC<PopupContentProps>;
+  Body: React.FC<PopupContentProps>;
+  Footer: React.FC<PopupContentProps>;
+};
