+++ types/Spinner.d.ts
@@ -1,0 +1,8 @@
+import React from "react";
+
+export type SpinnerProps = {
+  theme?: "dark" | "light";
+  className?: string;
+};
+
+export const Spinner: React.FC<SpinnerProps>;
